% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authenticate_sb.R
\name{authenticate_sb}
\alias{authenticate_sb}
\title{Authenticate to SB for subsequent calls}
\usage{
authenticate_sb(username, password)
}
\arguments{
\item{username}{Sciencebase username}

\item{password}{(optional) Sciencebase password, prompts user if not supplied}
}
\description{
This connects to SB, authenticates and gets a session token for
communicating with SB.
}
\details{

}

