% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/current_session.R
\name{current_session}
\alias{current_session}
\title{Return current cached session}
\usage{
current_session()
}
\description{
Returns the currently cached SB session. If there
is no authenticated session, returns NULL.
}
\examples{
session = current_session()
}
\author{
Luke Winslow
}

