% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/folder_create.R
\name{folder_create}
\alias{folder_create}
\title{Create a folder}
\usage{
folder_create(parent_id, name, ..., session = current_session())
}
\arguments{
\item{parent_id}{(character) the id of the parent folder/item under which to create this one}

\item{name}{(character) the folder name}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}} and
\code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
A \code{\link[httr]{response}} object
}
\description{
Create a folder
}
\examples{
\dontrun{
folder_create(user_id(), "foobar")
}
}

