% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_sb.R
\name{get_cached_token}
\alias{get_cached_token}
\title{get cached sciencebase token}
\usage{
get_cached_token()
}
\value{
character containing the token (which may be stale) or an empty string
}
\description{
tries to retrieve a cached token from `token_stache_path()`
}
