% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_item}
\alias{get_item}
\title{Retrieve SB item}
\usage{
get_item(id, session)
}
\arguments{
\item{id}{SB item ID}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
List serialization of complete metadata for SB item
}
\description{

}

