% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifier_exists.R
\name{identifier_exists}
\alias{identifier_exists}
\title{Check if identifier exists}
\usage{
identifier_exists(id, ..., session = current_session())
}
\arguments{
\item{id}{(character) A ScienceBase ID. See Details.}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{(optional) SB session from \link{authenticate_sb}}
}
\value{
Logical, \code{TRUE} or \code{FALSE}
}
\description{
Check if identifier exists
}
\details{
This function does not accept a sbitem class object because
we'd run \code{\link{as.sbitem}} on it, which would make a GET request,
and we only want to make a quick HEAD request.
}
\examples{
\dontrun{
# identifier exists
identifier_exists(id = "4f4e4b24e4b07f02db6aea14")

# identifier does not exist
identifier_exists(id = "aaaaaaakkkkkkkbbbbbb")
}
}

