% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_sb.R
\name{initialize_sciencebase_session}
\alias{initialize_sciencebase_session}
\title{Initialize ScienceBase Session}
\usage{
initialize_sciencebase_session(username, token_text = NULL)
}
\arguments{
\item{username}{email address of sciencebase user.}

\item{token_text}{character json formatted token text}
}
\description{
opens a browser for two factor authentication. A token can be
retrieved from the user drop down in the upper right once logged in.
The token should be pasted into the console. Can also be called with a pre-
fetched token.
}
