% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_upload_files.R
\name{item_append_check}
\alias{item_append_check}
\title{Upload File to Item With md5sum checks}
\usage{
item_append_check(
  sb_id,
  files,
  ...,
  title = rep("", length(files)),
  session = current_session()
)
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{files}{A string vector of paths to files to be uploaded}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{title}{character titles for appended files must be same length as files.}

\item{session}{Session object from \code{\link{authenticate_sb}}. Defaults to anonymous or 
last authenticated session}
}
\value{
An object of class \code{sbitem}
}
\description{
Adds a file to an item using the same
method as is implemente by the Web User Interface.
This method uses multiple web requests and may be
more fragile than the standard append functions.
}
\examples{
\dontrun{
res <- item_create(user_id(), "testing 123")
cat("foo bar", file = "foobar.txt")
item_append_check(res$id, "foobar.txt")
item_rm(res)
}
}
