% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item_exists}
\alias{item_exists}
\title{check if identifier touple already exists on SB}
\usage{
item_exists(scheme, type, key, session = NULL)
}
\arguments{
\item{scheme}{}

\item{type}{}

\item{key}{}

\item{session}{}
}
\value{
boolean for whether item exists
}
\description{
returns TRUE if touple already belongs to a sciencebase item, FALSE if not
}
\examples{
\dontrun{
item_exists('mda_streams','ts_doobs','nwis_01018035')
item_exists('mda_streams','site_root','nwis_01018035')
}
}

