% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_file_download.R
\name{item_file_download}
\alias{item_file_download}
\title{Download files attached to item}
\usage{
item_file_download(
  sb_id,
  ...,
  names,
  destinations,
  dest_dir = getwd(),
  overwrite_file = FALSE,
  fetch_cloud_urls = FALSE
)
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}.}

\item{names}{String vector list of file names attached to item that you wish
to download.}

\item{destinations}{String vector list of destinations for requested files.
Must be same length as \code{names}}

\item{dest_dir}{A directory path for saving files when \code{names} destinations 
parameter is not specified.}

\item{overwrite_file}{Boolean indicating if file should be overwritten if it
already exists locally}

\item{fetch_cloud_urls}{(logical) fetch a tokenized cloud download URLs? Default: \code{TRUE}
This option will take slightly longer but the `url` attribute of the returned list will
work for direct file downloads or use with other applications and libraries.}
}
\value{
Character vector of full paths to local files
}
\description{
Function to downlod files attached to an item on SB. Either files can be
specified directly using the \code{names} and \code{destinations} parameters,
or a \code{dest_dir} can be supplied where all attached files will be written
with the names as stored on SB.
}
\examples{
\dontrun{

#downloads all files attached to this item
item_file_download('627f1572d34e3bef0c9a30d8', dest_dir=tempdir())

#downloads a specific file attached to this item
item_file_download('627f1572d34e3bef0c9a30d8', names='example.txt',
		destinations=file.path(tempdir(), 'out.txt'))
		
}
}
