% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_item.R
\name{item_get}
\alias{item_get}
\title{Retrieve SB item}
\usage{
item_get(id, session)
}
\arguments{
\item{id}{SB item ID}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
List serialization of complete metadata for SB item
}
\description{

}

