% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_get_fields.R
\name{item_get_fields}
\alias{item_get_fields}
\title{Retrieve specific fields from an SB item}
\usage{
item_get_fields(id, fields, ..., drop = TRUE, session = current_session())
}
\arguments{
\item{id}{SB item ID}

\item{fields}{a vector of fields}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{drop}{logical. If only one field is selected, should the
list format be dropped?}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
List serialization of complete metadata for SB item
}
\description{
Retrieve specific fields from an SB item
}

