% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_get_wfs.R
\name{item_get_wfs}
\alias{item_get_wfs}
\title{Download and load from SB WFS service}
\usage{
item_get_wfs(sb_id, ..., session)
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}. Defaults to anonymous or 
last authenticated session}
}
\description{
This function attempts to download the spatial layer data attached to the
requested SB item. SB exposes discrete spatial objects (points, polygons) as 
WFS web services. This requires the GDAL libraries \code{rgdal} and \code{gdalUtils} 
to be installed, which may require the external GDAL libraries as well.
}
\examples{
\dontrun{
library(sp)
layer = item_get_wfs('55e372b9e4b05561fa208212')

plot(layer)

}

}

