% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_rm_recursive.R
\name{item_list_children_all}
\alias{item_list_children_all}
\title{Like item_list_children except that it returns all fields, not just 'id' and
'title'}
\usage{
item_list_children_all(id, ..., session = current_session(), limit = 100)
}
\arguments{
\item{id}{SB item ID}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{(optional) SB session from \link{authenticate_sb}}

\item{limit}{Max children returned}
}
\description{
Like item_list_children except that it returns all fields, not just 'id' and
'title'
}
\keyword{internal}

