% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_rm_files.R
\name{item_rm_files}
\alias{item_rm_files}
\title{Remove all files associated with an item}
\usage{
item_rm_files(id, ..., session = current_session())
}
\arguments{
\item{id}{a sciencebase item identifier}

\item{...}{Additional parameters are passed on to \code{\link[httr]{PUT}}}

\item{session}{Authenticated session object (from \link{authenticate_sb})}
}
\description{
Removes existing files associated with an item.
}

