% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_update_identifier.R
\name{item_update_identifier}
\alias{item_update_identifier}
\title{Add custom identifier to an existing item}
\usage{
item_update_identifier(id, scheme, type, key, ...,
  session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{scheme}{The identifier scheme}

\item{type}{The identifier type}

\item{key}{The identifier key}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\examples{

\dontrun{

session = authenticate_sb("user@usgs.gov")
item_update_identifier("5485fd99e4b02acb4f0c7e81", "scheme", "type", "key", session=session)

}

}
\author{
Luke Winslow
}

