% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_get_wfs.R
\name{match_url_distro}
\alias{match_url_distro}
\title{check that it has a WFS service, throw error if missing}
\usage{
match_url_distro(item, title_to_match)
}
\description{
check that it has a WFS service, throw error if missing
}

