% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query_item_identifier}
\alias{query_item_identifier}
\title{Query SB for item based on custom identifier}
\usage{
query_item_identifier(scheme, type, key, session)
}
\arguments{
\item{scheme}{The identifier scheme}

\item{type}{The identifier type}

\item{key}{The identifier key}
}
\value{
The SB item id for the matching item
}
\description{

}

