% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_item_in_folder.R
\name{query_item_in_folder}
\alias{query_item_in_folder}
\title{Search within an SB folder}
\usage{
query_item_in_folder(text, folder, ..., limit = 20)
}
\arguments{
\item{text}{text in the title, abstract, etc. of the desired item}

\item{folder}{an SB item ID for the folder to search in}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{limit}{Max number of matching items to return}

\item{session}{(optional) SB Session to use, not provided queries public 
items only}
}
\value{
A list of matching items as sbitem objects.
}
\description{
Search for text in the title, abstract, etc. within an SB folder and any
subfolders.
}
