% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query_items.R
\name{query_items}
\alias{query_items}
\title{Query SB for items using generic query parameters}
\usage{
query_items(query_list, session)
}
\arguments{
\item{query_list}{}

\item{session}{Session object from \code{\link{authenticate_sb}}}

\item{query_list}{List of item query selectors}
}
\description{

}

