% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\name{sb_ping}
\alias{sb_ping}
\title{Ping ScienceBase to see if it's available}
\usage{
sb_ping(...)
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}
}
\value{
Boolean (TRUE) indicating if a connection to ScienceBase can be established 
and if it is responding as expected. FALSE otherwise.
}
\description{
Ping ScienceBase to see if it's available
}
\examples{
\donttest{
#TRUE if all is well and SB can be contacted
sb_ping()
}
}
