% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST_helpers.R
\name{sbtools_DELETE}
\alias{sbtools_DELETE}
\title{generic DELETEs for sbtools package}
\usage{
sbtools_DELETE(url, ...)
}
\arguments{
\item{url}{a base url for the DELETE}

\item{...}{additional params passed to \code{\link[httr]{DELETE}}, e.g., 
\code{accept_json()}}
}
\description{
package wrapped for generic DELETEs that test sessions internally and wrap
some errors
}
\keyword{internal}
