% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REST_helpers.R
\name{sbtools_GET}
\alias{sbtools_GET}
\title{generic GETs for sbtools package
package wrapped for generic GETs that test sessions internally and wrap some errors}
\usage{
sbtools_GET(url, query, session, timeout = default_timeout(), ...)
}
\arguments{
\item{url}{a base url for the GET}

\item{query}{a list of query params}

\item{session}{a sbtools session object}

\item{timeout}{a timeout for the request in seconds}

\item{...}{additional params passed to \code{\link[httr]{GET}}}
}
\description{
generic GETs for sbtools package
package wrapped for generic GETs that test sessions internally and wrap some errors
}
\keyword{internal}

