% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST_helpers.R
\name{sbtools_POST}
\alias{sbtools_POST}
\title{generic POSTs for sbtools package
package wrapped for generic POSTs that test sessions internally and wrap some errors}
\usage{
sbtools_POST(url, body, ..., session)
}
\arguments{
\item{url}{a base url for the POST}

\item{body}{a POST body}

\item{...}{additional params passed to \code{\link[httr]{POST}}}

\item{session}{a sbtools session object}
}
\description{
generic POSTs for sbtools package
package wrapped for generic POSTs that test sessions internally and wrap some errors
}
\keyword{internal}
