% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST_helpers.R
\name{sbtools_PUT}
\alias{sbtools_PUT}
\title{generic PUTs for sbtools package}
\usage{
sbtools_PUT(url, body, ..., session)
}
\arguments{
\item{url}{a base url for the PUT}

\item{body}{the PUT body as in \code{\link[httr]{PUT}}}

\item{...}{additional params passed to \code{\link[httr]{PUT}}, e.g.,
\code{accept_json()}}

\item{session}{a sbtools session object}
}
\description{
package wrapped for generic PUTs that test sessions internally and wrap some 
errors
}
\keyword{internal}
