% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session_age.R
\name{session_age}
\alias{session_age}
\title{a convienence function for getting the age of a session.}
\usage{
session_age(session = current_session())
}
\arguments{
\item{session}{an SB session}
}
\value{
difftime object
}
\description{
See pkg.env$session_birthdate for time when session was created.
See pkg.env$session_expires for session age when session goes stale
}
\examples{
\dontrun{
authenticate_sb('bbadger@usgs.gov')
sbtools::session_age()
}
}
\keyword{internal}

