% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_age_reset.R
\name{session_age_reset}
\alias{session_age_reset}
\title{session reset age}
\usage{
session_age_reset(session)
}
\arguments{
\item{session}{session object (from \link{authenticate_sb})}
}
\description{
internal function for refreshing session age to 0 using Sys.time()
}
\keyword{internal}
