% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_authorized.R
\name{session_authorized}
\alias{session_authorized}
\title{Determine whether a session is authenticated}
\usage{
session_authorized(session = current_session())
}
\arguments{
\item{session}{The session to check}
}
\description{
Returns TRUE if the session is a valid, non-NULL, active SB session.
}

