% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_check_reauth.R
\name{session_check_reauth}
\alias{session_check_reauth}
\title{Checks current session and re-authenticates if necessary}
\usage{
session_check_reauth(session, username, password, ...)
}
\arguments{
\item{session}{(Advanced) Alternatively supply a specific session to be checked}

\item{username}{Sciencebase username}

\item{password}{Sciencebase password, prompts user if not supplied}

\item{...}{Any additional parameters are ignored.}
}
\description{
This checks the state of your Sciencebase session and, if invalid or if you've supplied
a new user name, it will re-authenticate. This is primarily a helper function for 
packages that depend on sbtools.
}
\examples{

\dontrun{

#will prompt for username/pass if needed
session_check_reauth()

session_check_reauth('user@usgs.gov')

#default behavior
session_check_reauth(session=current_session())

}


}

