% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session_age.R
\name{session_expired}
\alias{session_expired}
\title{Check whether an SB session is expired}
\usage{
session_expired(session = current_session())
}
\arguments{
\item{session}{the SB session}
}
\description{
Check the expiration using session_age
}
\keyword{internal}

