% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session_info.R
\name{session_info}
\alias{session_info}
\title{Get session info}
\usage{
session_info(..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}
}
\value{
list, if not logged in states that, but if logged in, user details
}
\description{
Get session info
}
\examples{
\dontrun{
session_info()
}
}

