% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_logout.R
\name{session_logout}
\alias{session_logout}
\title{Logout of a ScienceBase session}
\usage{
session_logout(..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}
}
\value{
invisible, returns nothing if logged out, or errors with message
}
\description{
Logout of a ScienceBase session
}
\examples{
\dontrun{
session_logout()
}
}
