% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session_set.R
\name{session_set}
\alias{session_set}
\title{Set the session to a user-specified value}
\usage{
session_set(session = NULL)
}
\arguments{
\item{session}{The new session, or NULL}
}
\description{
Set the session to a user-specified value
}
\keyword{internal}

