% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_endpoint.R
\name{set_endpoint}
\alias{set_endpoint}
\title{Set SB endpoint}
\usage{
set_endpoint(endpoint = "production")
}
\arguments{
\item{endpoint}{Indicate which SB endpoint
you want to use options: \code{c('production','development')}}
}
\description{
Sets the internal URLS used to either the production
or development (beta) SB server. URLS are stored internally to the package
}
\examples{
\dontrun{
set_endpoint('prod')

# getting item from production SB servers
item_get('54b94448e4b0ed73b67550ea')

set_endpoint('dev')
# getting item from beta SB servers
item_get('54b94448e4b0ed73b67550ea')

}
}
\author{
Luke Winslow
}

