% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AAA.R
\name{set_expiration}
\alias{set_expiration}
\title{Set the expiration}
\usage{
set_expiration(expiration_age = as.difftime("00:59:00"))
}
\arguments{
\item{expiration_age}{a difftime describing the length of a session}
}
\description{
Called on package build
}
\keyword{internal}

