% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\name{user_id}
\alias{user_id}
\title{Get your parent ID}
\usage{
user_id(...)
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{POST}}}
}
\value{
A single character string, your user id
}
\description{
Required for creating items
}
\examples{
\dontrun{
user_id()
}
}
