% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_create.R
\name{folder_create}
\alias{folder_create}
\title{Create a folder}
\usage{
folder_create(sb_id, name, ..., session = current_session())
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{name}{(character) the folder name}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}. Defaults to anonymous or 
last authenticated session}

\item{sb_parent}{An \code{\link{sbitem}} object or character ScienceBase ID corresponding to the 
parent item (folder)}
}
\value{
A \code{\link[httr]{response}} object
}
\description{
Create a folder
}
\examples{
\dontrun{
folder_create(user_id(), "foobar345")
}
}

