% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_create.R
\name{folder_create}
\alias{folder_create}
\title{Create a folder}
\usage{
folder_create(id, name, ..., session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{name}{(character) the folder name}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
A \code{\link[httr]{response}} object
}
\description{
Create a folder
}
\examples{
\dontrun{
folder_create(user_id(), "foobar345")
}
}

