% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_sb.R
\name{get_username}
\alias{get_username}
\title{Get or set ScienceBase username}
\usage{
get_username(username = NULL)
}
\arguments{
\item{username}{character if NULL, will be retrieved from the `sb_user`
environment variable or the `username` file stored at `token_stache_path()`}
}
\value{
character string containing a username. Throws an error if no
username is found and `interactive()` is FALSE
}
\description{
Used to retrieve the current user name. Will request the username be entered
if no user is provided and `interactive()` is TRUE.

This is largely an internal function, but is exported for awareness and
use in automated pipelines.
}
