% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_exists.R
\name{identifier_exists}
\alias{identifier_exists}
\title{Check if identifier exists}
\usage{
identifier_exists(sb_id, ..., session = current_session())
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}. Defaults to anonymous or 
last authenticated session}

\item{sb_parent}{An \code{\link{sbitem}} object or character ScienceBase ID corresponding to the 
parent item (folder)}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{(optional) SB session from \link{authenticate_sb}}
}
\value{
Logical, \code{TRUE} or \code{FALSE}
}
\details{
This function does not accept a sbitem class object because 
we'd run \code{\link{as.sbitem}} on it, which would make a GET request,
and we only want to make a quick HEAD request.
}
\examples{
\dontrun{
# identifier exists
identifier_exists(id = "4f4e4b24e4b07f02db6aea14")

# identifier does not exist
identifier_exists(id = "aaaaaaakkkkkkkbbbbbb")
}
}

