% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_logged_in.R
\name{is_logged_in}
\alias{is_logged_in}
\title{Check whether you're logged into a ScienceBase session}
\usage{
is_logged_in(..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}
}
\value{
Logical, \code{TRUE} or \code{FALSE}
}
\description{
Check whether you're logged into a ScienceBase session
}
\examples{
\dontrun{
is_logged_in()
}
}
