% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item_create}
\alias{item_create}
\title{Create a new SB item}
\usage{
item_create(parent_id, title, session, info)
}
\arguments{
\item{parent_id}{ID of the parent item}

\item{session}{Authenticated session object (from \link{authenticate_sb})}

\item{info}{(optional) list of metadata info for the new item}
}
\value{
SB ID for the newly created item
}
\description{
Create a new SB item
}

