% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create.R
\name{item_create}
\alias{item_create}
\title{Create a new SB item}
\usage{
item_create(id = user_id(), title, ..., info, session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{title}{The title of the new SB item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{info}{(optional) list of metadata info for the new item}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
An object of class \code{sbitem}
}
\examples{
\dontrun{
# Create an item - by default we use your user ID
item_create(title = "testing 123")

# Pass an object of class sbitem
x <- folder_create(user_id(), "foobar456")
item_create(x, "foobar456-item")
}
}

