% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_exists.R
\name{item_exists}
\alias{item_exists}
\title{check if identifier touple already exists on SB}
\usage{
item_exists(scheme, type, key, ..., session = current_session())
}
\arguments{
\item{scheme}{the identifier scheme}

\item{type}{the identifier type}

\item{key}{the identifier key}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{an SB session}
}
\value{
boolean for whether item exists
}
\description{
returns TRUE if touple already belongs to a sciencebase item, FALSE if not
}
\examples{
\dontrun{
item_exists('mda_streams','ts_doobs','nwis_01018035')
item_exists('mda_streams','site_root','nwis_01018035')
}
}
