% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_get.R
\name{item_get}
\alias{item_get}
\title{Retrieve SB item}
\usage{
item_get(id, ..., session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
An object of class \code{sbitem}
}
\description{
Retrieve SB item
}
\examples{
\dontrun{
# Get an item
item_get("4f4e4b24e4b07f02db6aea14")

# Search for item IDs, then pass to item_get
library("httr")
res <- query_items(list(s = "Search", q = "water", format = "json"))
ids <- vapply(httr::content(res)$items, "[[", "", "id")
lapply(ids[1:3], item_get)
}
}

