% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_get_parent.R
\name{item_get_parent}
\alias{item_get_parent}
\title{Get an item's parent ID}
\usage{
item_get_parent(sbitem, ..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}

\item{id}{ID of SB item}
}
\value{
The parent ID of an item.
}
\description{
Get an item's parent ID
}
\examples{
\dontrun{
item_get_parent("4f4e4b24e4b07f02db6aea14")

item_get("4f4e4b24e4b07f02db6aea14") \%>\% item_get_parent()
}
}

