% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_get_parent.R
\name{item_get_parent}
\alias{item_get_parent}
\title{Get an item's parent ID}
\usage{
item_get_parent(id, ..., session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
The parent ID of an item.
}
\examples{
\dontrun{
item_get_parent("4f4e4b24e4b07f02db6aea14")

item_get("4f4e4b24e4b07f02db6aea14") \%>\% item_get_parent()
}
}

