% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_get_parent.R
\name{item_get_parent}
\alias{item_get_parent}
\title{Get an item's parent ID}
\usage{
item_get_parent(id, ..., session = current_session())
}
\arguments{
\item{id}{ID of SB item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}
}
\description{
Returns the parent ID of an item.
}

