% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_list_children.R
\name{item_list_children}
\alias{item_list_children}
\title{Return IDs for all child items}
\usage{
item_list_children(id, ..., session = current_session(), limit = 20)
}
\arguments{
\item{id}{SB item ID}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{(optional) SB session from \link{authenticate_sb}}

\item{limit}{Max children returned}
}
\value{
\code{data.frame} with a row for each child item
}
\description{
Returns a list of child IDs for a ScienceBase item
}
\examples{
item_list_children('5060b03ae4b00fc20c4f3c8b')
}

