% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_list_children.R
\name{item_list_children}
\alias{item_list_children}
\title{Return IDs for all child items}
\usage{
item_list_children(sb_id, ..., session = current_session(), limit = 20)
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}. Defaults to anonymous or 
last authenticated session}

\item{limit}{Max children returned}

\item{sb_parent}{An \code{\link{sbitem}} object or character ScienceBase ID corresponding to the 
parent item (folder)}
}
\value{
\code{data.frame} with a row for each child item
}
\description{
Returns a list of child IDs for a ScienceBase item
}
\examples{
\dontrun{
item_list_children('5060b03ae4b00fc20c4f3c8b')

as.sbitem('5060b03ae4b00fc20c4f3c8b') \%>\% item_list_children
item_get('5060b03ae4b00fc20c4f3c8b') \%>\% item_list_children
}
}

