% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item_list_files}
\alias{item_list_files}
\title{Get list of files attached to SB item}
\usage{
item_list_files(id, session)
}
\arguments{
\item{SB}{item ID}

\item{Session}{from \link{authenticate_sb}}
}
\description{
Get list of files attached to SB item
}

