% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item_rm}
\alias{item_rm}
\title{Remove item from SB}
\usage{
item_rm(id, session)
}
\arguments{
\item{session}{Session object from \code{\link{authenticate_sb}}}

\item{SB}{item ID}
}
\value{
TRUE to indicate success
}
\description{
Remove item from SB
}

