% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item_rm.R
\name{item_rm}
\alias{item_rm}
\title{Remove item from SB}
\usage{
item_rm(id, session)
}
\arguments{
\item{id}{item ID}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
TRUE to indicate success
}
\description{
Remove item from SB
}

