% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_rm.R
\name{item_rm}
\alias{item_rm}
\title{Remove item from SB}
\usage{
item_rm(id, ..., session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID
by \code{\link{as.sbitem}}}

\item{...}{Additional parameters are passed on to \code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
\pkg{httr} \code{\link[httr]{response}} object
}
\description{
Remove item from SB
}
\examples{
\dontrun{
res <- item_create(user_id(), "item-to-delete")
item_rm(res)
}
}

