% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_remove_files.R
\name{item_rm_files}
\alias{item_rm_files}
\title{Remove all files associated with an item}
\usage{
item_rm_files(item_id, session)
}
\arguments{
\item{item_id}{a sciencebase item identifier}

\item{session}{Authenticated session object (from \link{authenticate_sb})}
}
\description{
Removes existing files associated with an item.
}

