% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_rm_recursive.R
\name{item_rm_recursive}
\alias{item_rm_recursive}
\title{Remove an item completely by recursively removing its child items}
\usage{
item_rm_recursive(id, ..., session = current_session())
}
\arguments{
\item{id}{A ScienceBase ID or something that can be coerced to a SB item ID}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}} and
\code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
\code{TRUE} to indicate success
}
\description{
Remove an item completely by recursively removing its child items
}
\details{
BEWARE: This completely removes ALL CHILD ITEMS AND THEIR CHILDREN
  as well as the item itself.
}
\examples{
\dontrun{
# Create an item with nested structure
authenticate_sb()
fname <- "chairs"
folder = folder_create(user_id(), name = fname)
item_create(folder, title='child1')
item_create(folder, title='child2')

item_list_children(folder)

# then delete the whole folder
sbtools:::item_rm_recursive(folder)
}
}
\keyword{internal}

