% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_rm_recursive.R
\name{item_rm_recursive}
\alias{item_rm_recursive}
\title{Remove an item completely by recursively removing its children}
\usage{
item_rm_recursive(id, ..., session = current_session())
}
\arguments{
\item{id}{(character) item ID}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}} and
\code{\link[httr]{DELETE}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\value{
\code{TRUE} to indicate success
}
\description{
Remove an item completely by recursively removing its children
}
\details{
BEWARE: This removes all folders/files in an item completely.
}
\examples{
\dontrun{
# Create an item with nested structure
fname <- "chairs"
folder_create(name = fname)
df <- item_list_children(user_id())
id <- df[ df$title == fname, "id" ]
folder_create(id, name = "one")
folder_create(id, name = "two")
df2 <- item_list_children(id)
folder_create(df2$id[1], name = "nested")

# then delete the whole folder
item_rm_recursive(id)
}
}

