% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_update.R
\name{item_update}
\alias{item_update}
\title{Update a SB item with new metadata}
\usage{
item_update(id, info, ..., session = current_session())
}
\arguments{
\item{id}{SB item ID}

\item{info}{list of metadata info for the new item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{PUT}}}

\item{session}{Authenticated session object (from \link{authenticate_sb})}
}
\value{
List serialization of complete metadata for SB item
}
\description{
Update a SB item with new metadata
}

