% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item_update_identifier}
\alias{item_update_identifier}
\title{Add custom identifier to an existing item}
\usage{
item_update_identifier(id, scheme, type, key, session)
}
\arguments{
\item{id}{SB item ID}

\item{scheme}{The identifier scheme}

\item{type}{The identifier type}

\item{key}{The identifier key}

\item{session}{}
}
\description{
Add custom identifier to an existing item
}

