% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_update_identifier.R
\name{item_update_identifier}
\alias{item_update_identifier}
\title{Add custom identifier to an existing item}
\usage{
item_update_identifier(id, scheme, type, key, session = current_session())
}
\arguments{
\item{id}{SB item ID}

\item{scheme}{The identifier scheme}

\item{type}{The identifier type}

\item{key}{The identifier key}

\item{session}{SB session from \code{\link{authenticate_sb}}}
}
\description{
Add custom identifier to an existing item
}
\examples{
\dontrun{

session = authenticate_sb("user@usgs.gov")
item_update_identifier("5485fd99e4b02acb4f0c7e81", "scheme", "type", "key", session)

}
}
\author{
Luke Winslow
}

