% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_upload_files.R
\name{item_upload_create}
\alias{item_upload_create}
\title{#' Upload file(s) and create a new item}
\usage{
item_upload_create(parent_id, files, ..., scrape_files = TRUE)
}
\arguments{
\item{parent_id}{An \code{\link{sbitem}} object or character ScienceBase ID corresponding to the 
parent item (folder)}

\item{files}{A string vector of paths to files to be uploaded}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}}

\item{scrape_files}{logical should the files be scraped for metadata? 
If TRUE, sciencebase will attempt to create extensions based on the files.

For example, for shapefiles, this will result in a shapefile extension 
to be returned as a facet of the sciencebase item. 
See item: "58069258e4b0824b2d1d422e" for an example.}
}
\value{
An object of class \code{sbitem}
}
\description{
Create a new item with files attached, all in one call to SB
}
\examples{
\dontrun{
# You'll need a parent id for a folder/item
## here, using your highest level parent folder
file <- system.file("examples", "books.json", package = "sbtools")
item_upload_create(user_id(), file)
}
}
