% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_upload_create.R
\name{item_upload_create}
\alias{item_upload_create}
\title{Upload file(s) and create a new item}
\usage{
item_upload_create(parent_id, files, session)
}
\arguments{
\item{parent_id}{SB id of parent item for the newly created item}

\item{files}{A string vector of paths to files to be uploaded}

\item{session}{Session object from authenticate_sb}
}
\description{

}

