% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/item_upload_create.R
\name{item_upload_create}
\alias{item_upload_create}
\title{Upload file(s) and create a new item}
\usage{
item_upload_create(parent_id, files, ..., session = current_session())
}
\arguments{
\item{parent_id}{SB id of parent item for the newly created item}

\item{files}{A string vector of paths to files to be uploaded}

\item{...}{Additional parameters are passed on to \code{\link[httr]{POST}}}

\item{session}{Session object from authenticate_sb}
}
\description{
Create a new item with files attached, all in one call to SB
}
\examples{
\dontrun{
# You'll need a parent id for a folder/item
### these examples aren't run on check b/c specific to a user account
# file <- system.file("examples", "books.json", package = "sbtools")
# item_upload_create("55ddf814e4b0518e354dfcb4", file)
}
}

