% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_create.R
\name{items_create}
\alias{items_create}
\title{Create many new SB items}
\usage{
items_create(parent_id = user_id(), title, ..., info = NULL)
}
\arguments{
\item{parent_id}{An \code{\link{sbitem}} object or character ScienceBase ID 
corresponding to the parent item (folder). This must be of length 1 or more. 
If length 1, then we recycle it for every item.}

\item{title}{Two or more titles for the new SB items}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{info}{(optional) list of metadata info for the new items. for each 
item include a named list of variables}
}
\value{
One or more objects of class \code{sbitem} in a list
}
\description{
A method to create multiple ScienceBase items with a single call and a single HTTP service
request. Can be useful for improving performance of creating a large number of items at once.
}
\details{
The length of the \code{title} and \code{info} values must be the same
length - however, the \code{parent_id} can be of length 1 or equal to the length 
of each of \code{title} and \code{info} parameters
}
\examples{
\dontrun{
# helper function to make a random name
aname <- function() paste0(sample(letters, size = 5, replace = TRUE), collapse = "")

# Create some items - by default we use your user ID
items_create(title = c(aname(), aname()))

# add additional items in the info parameter - by default we use your user ID
items_create(title = c(aname(), aname()), 
		info = list(
						list(contacts = list(list(name = "Suzy"))), 
						list(contacts = list(list(name = "Brandy")))
					)
	)
	
# another example with more information - by default we use your user ID
items_create(title = c(aname(), aname()), 
		info = list(
						list(contacts = list(list(name = "Suzy"))), 
						list(contacts = list(list(name = "Brandy")))
					)
	)	

# Pass an object of class sbitem
(x <- folder_create(user_id(), aname()))
items_create(x, title = c(aname(), aname()))
}
}
