% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query_item_identifier}
\alias{query_item_identifier}
\title{Query SB for items based on custom identifier}
\usage{
query_item_identifier(scheme, type, key, session, limit = 20)
}
\arguments{
\item{scheme}{The identifier scheme}

\item{type}{(optional) The identifier type}

\item{key}{(optional) The identifier key}

\item{session}{(optional) SB Session to use, not provided queries public items only}

\item{limit}{Max number of matching items to return}
}
\value{
The SB item id for the matching item. NULL if no matching item found.
}
\description{

}

