% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sb_spatial.R
\name{query_sb_spatial}
\alias{query_sb_spatial}
\title{Query SB based on spatial extent}
\usage{
query_sb_spatial(bbox, bb_min, bb_max, ..., limit = 20,
  session = current_session())
}
\arguments{
\item{bbox}{An object of class \code{\link{sp}}. The bounding box of the object is used for the query.}

\item{bb_min}{Alternate to bbox param. A vector length 2 defining the c(long, lat) of lower left corner of the bounding box.}

\item{bb_max}{Alternate to bbox param. A vector length 2 defining the c(long, lat) of upper right corner of the bounding box.}
}
\description{
Queries ScienceBase based on a spatial bounding box. Accepts either an object of class \code{\link{sp}} 
(uses the spatial object's bounding box) or two long/lat coordinates defining the lower left and upper 
right corners.
}
\examples{

query_sb_spatial(bb_min=c(-104.4, 37.5), bb_max=c(-95.1, 41.0))


}

