% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sb_ping.R
\name{sb_ping}
\alias{sb_ping}
\title{Ping ScienceBase to see if it's available}
\usage{
sb_ping(...)
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}
}
\value{
list, with \code{"OK"} if ScienceBase is up
}
\description{
Ping ScienceBase to see if it's available
}
\examples{
\dontrun{
sb_ping()
}
}

