% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sb_item.R
\name{sbitem}
\alias{as.sbitem}
\alias{is.sbitem}
\alias{sbitem}
\title{ScienceBase item class}
\usage{
as.sbitem(x, ...)

is.sbitem(x)
}
\description{
ScienceBase item class
}
\examples{
\dontrun{
# Single item from item_get()
item_get("4f4e4b24e4b07f02db6aea14")

# Get many w/ e.g., an lapply() call
library("httr")
res <- query_items(list(s = "Search", q = "water", format = "json"))
ids <- vapply(httr::content(res)$items, "[[", "", "id")
(out <- lapply(ids[1:3], item_get))

# create item class from only an item ID
as.sbitem("4f4e4b24e4b07f02db6aea14")

# sbitem gives back itself
(x <- as.sbitem("4f4e4b24e4b07f02db6aea14"))
as.sbitem(x)
}
}

