% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbtools-package.R
\docType{package}
\name{sbtools-package}
\alias{sbtools}
\alias{sbtools-package}
\title{R interface to ScienceBase}
\description{
This package provides a rich interface to USGS's ScienceBase
\url{https://www.sciencebase.gov/} - a data cataloging and collaborative
data management platform.

Functions are inluded for searching for data, retrieving, creating, and
updating datasets.
}
\section{Authentication}{


See the function \code{\link{authenticate_sb}} to authenticate. You'll be
required to pass in your ScienceBase username and password.
}

\section{Feedback}{


Report any feedback or bugs at \url{https://github.com/USGS-R/sbtools/issues}
}
\keyword{package}

