% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST_helpers.R
\name{sbtools_GET}
\alias{sbtools_GET}
\title{generic GETs for sbtools package}
\usage{
sbtools_GET(url, ..., session)
}
\arguments{
\item{url}{a base url for the GET}

\item{...}{additional params passed to \code{\link[httr]{GET}}, often 
including \code{query}}

\item{session}{a sbtools session object}
}
\description{
package wrapped for generic GETs that test sessions internally and wrap some
errors
}
\keyword{internal}

