% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_session.R
\name{session_age}
\alias{session_age}
\title{a convienence function for getting the age of a session.}
\usage{
session_age()
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}.}
}
\value{
difftime object
}
\description{
a convienence function for getting the age of a session.
}
\examples{
\dontrun{
authenticate_sb('bbadger@usgs.gov')
sbtools::session_age()
}
}
\keyword{internal}
