% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_details.R
\name{session_details}
\alias{session_details}
\title{Get session info}
\usage{
session_details(..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}}

\item{session}{SB session object from \code{\link{authenticate_sb}}}
}
\value{
list, if not logged in states that, but if logged in, user details
}
\examples{
\dontrun{
session_info()
}
}

