% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_session.R
\name{session_details}
\alias{session_details}
\title{Get session info (deprecated)}
\usage{
session_details()
}
\value{
list, if not logged in states that, but if logged in, user details
}
\description{
Get the details associated with current ScienceBase user session.
}
\examples{
\dontrun{

session_details()

}
}
