% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_session.R
\name{session_expired}
\alias{session_expired}
\title{Check whether an SB session is expired}
\usage{
session_expired()
}
\arguments{
\item{sb_id}{An \code{\link{sbitem}} object or a character ScienceBase ID corresponding to the item}

\item{...}{Additional parameters are passed on to \code{\link[httr]{GET}}, \code{\link[httr]{POST}},
\code{\link[httr]{HEAD}}, \code{\link[httr]{PUT}}.}
}
\description{
Check the expiration using session_age
}
\keyword{internal}
