% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_session.R
\name{session_validate}
\alias{session_validate}
\title{Validate sbtools session state}
\usage{
session_validate()
}
\value{
\code{TRUE/FALSE} indicating if session is valid and can be used. 
  Returns TRUE if session is NULL as well.
}
\description{
A session is considered valid if it is NULL or a true, non-expired SB session
}
\details{
This function only operates on the active initialized session.
}
