% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_endpoint.R
\name{set_endpoint}
\alias{set_endpoint}
\title{Set SB endpoint}
\usage{
set_endpoint(endpoint = c("production", "development"))
}
\arguments{
\item{endpoint}{Indicate which SB endpoint 
you want to use options: \code{c('production','development')}}
}
\description{
Sets the internal URLS used to either the production
or development (beta) SB server. URLS are stored internally to the package
}
\examples{
\donttest{
set_endpoint('prod')

# getting item from production SB servers
item_get('5060b03ae4b00fc20c4f3c8b')

set_endpoint('dev')
# getting item from beta SB servers
item_get('521e4686e4b051c878dc35d0')

}

}
\author{
Luke Winslow
}
