% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_id.R
\name{user_id}
\alias{user_id}
\title{Get your parent ID}
\usage{
user_id(..., session = current_session())
}
\arguments{
\item{...}{Additional parameters are passed on to \code{\link[httr]{POST}}}

\item{session}{Session object from \code{\link{authenticate_sb}}}
}
\description{
Required for creating items
}
\examples{
\dontrun{
user_id()
}
}

